requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.4.0
gunicorn==22.0.0
whitenoise==6.7.0
xhtml2pdf==0.2.15
openpyxl==3.1.5
drf-spectacular==0.27.2
redis==5.0.1
django-redis==5.4.0
prometheus-client==0.19.0
django-prometheus==2.3.1
psutil==5.9.8
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0


